+++ CHANGELOG.md
@@ -5,3 +5,10 @@
     - fín    [17/05/2022]
     - link Curso: https://docs.microsoft.com/en-us/learn/browse/?terms=typescript
     - link GitHub: https://github.com/CesarRomeroDev/TypeScript
+
+## [Declarar tipos de variables en TypeScript]
+    - inicio [18/05/2022]
+    - Seguimiento
+    - Secciónes:
+        - Interferencia de Tipos
+        - Enumeraciones